+++ Sweet.Redis.v2/Common/Result/RedisBool.cs
@@ -200,7 +200,7 @@ namespace Sweet.Redis.v2
         {
             if (ReferenceEquals(a, null))
                 return false;
-            return (b.m_Status == (long)RedisResultStatus.Completed) && ((bool)b.m_RawData == a);
+            return (b.m_Status == (int)RedisResultStatus.Completed) && ((bool)b.m_RawData == a);
         }
 
         public static bool operator !=(bool a, RedisBool b)
@@ -212,7 +212,7 @@ namespace Sweet.Redis.v2
         {
             if (ReferenceEquals(a, null))
                 return false;
-            return (a.m_Status == (long)RedisResultStatus.Completed) && ((bool)a.m_RawData == b);
+            return (a.m_Status == (int)RedisResultStatus.Completed) && ((bool)a.m_RawData == b);
         }
 
         public static bool operator !=(RedisBool a, bool b)
